+++ app/src/main/java/com/flint/core/designsystem/component/button/FlintModalButton.kt
@@ -1,4 +1,4 @@
-package com.flint.core.designsystem.component.modal
+package com.flint.core.designsystem.component.button
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
